refactor: modularize bootstrap option sets
