+++ data/bootstrap-classes/align-items-options.js
@@ -1,0 +1,27 @@
+export const alignItemsOptions = [
+	{ label: 'Align Items Start', value: 'align-items-start' },
+	{ label: 'Medium Screen Align Start', value: 'align-items-md-start' },
+	{ label: 'Large Screen Align Start', value: 'align-items-lg-start' },
+	{ label: 'Extra Large Align Start', value: 'align-items-xl-start' },
+	{ label: 'XXL Align Start', value: 'align-items-xxl-start' },
+	{ label: 'Align Items Center', value: 'align-items-center' },
+	{ label: 'Medium Screen Align Center', value: 'align-items-md-center' },
+	{ label: 'Large Screen Align Center', value: 'align-items-lg-center' },
+	{ label: 'Extra Large Align Center', value: 'align-items-xl-center' },
+	{ label: 'XXL Align Center', value: 'align-items-xxl-center' },
+	{ label: 'Align Items End', value: 'align-items-end' },
+	{ label: 'Medium Screen Align End', value: 'align-items-md-end' },
+	{ label: 'Large Screen Align End', value: 'align-items-lg-end' },
+	{ label: 'Extra Large Align End', value: 'align-items-xl-end' },
+	{ label: 'XXL Align End', value: 'align-items-xxl-end' },
+	{ label: 'Align Items Baseline', value: 'align-items-baseline' },
+	{ label: 'Medium Screen Align Baseline', value: 'align-items-md-baseline' },
+	{ label: 'Large Screen Align Baseline', value: 'align-items-lg-baseline' },
+	{ label: 'Extra Large Align Baseline', value: 'align-items-xl-baseline' },
+	{ label: 'XXL Align Baseline', value: 'align-items-xxl-baseline' },
+	{ label: 'Align Items Stretch', value: 'align-items-stretch' },
+	{ label: 'Medium Screen Align Stretch', value: 'align-items-md-stretch' },
+	{ label: 'Large Screen Align Stretch', value: 'align-items-lg-stretch' },
+	{ label: 'Extra Large Align Stretch', value: 'align-items-xl-stretch' },
+	{ label: 'XXL Align Stretch', value: 'align-items-xxl-stretch' },
+];
