+++ data/bootstrap-classes/bleed-cover-options.js
@@ -1,0 +1,10 @@
+export const bleedCoverOptions = [
+	{
+		label: 'Extend Cover Left',
+		value: 'cover-negative-margin-left',
+	},
+	{
+		label: 'Extend Cover Right',
+		value: 'cover-negative-margin-right',
+	},
+];
