+++ data/bootstrap-classes/blend-mode-options.js
@@ -1,0 +1,18 @@
+export const blendModeOptions = [
+	{ label: 'Normal', value: 'mix-blend-normal' },
+	{ label: 'Multiply', value: 'mix-blend-multiply' },
+	{ label: 'Screen', value: 'mix-blend-screen' },
+	{ label: 'Overlay', value: 'mix-blend-overlay' },
+	{ label: 'Darken', value: 'mix-blend-darken' },
+	{ label: 'Lighten', value: 'mix-blend-lighten' },
+	{ label: 'Color Dodge', value: 'mix-blend-color-dodge' },
+	{ label: 'Color Burn', value: 'mix-blend-color-burn' },
+	{ label: 'Hard Light', value: 'mix-blend-hard-light' },
+	{ label: 'Soft Light', value: 'mix-blend-soft-light' },
+	{ label: 'Difference', value: 'mix-blend-difference' },
+	{ label: 'Exclusion', value: 'mix-blend-exclusion' },
+	{ label: 'Hue', value: 'mix-blend-hue' },
+	{ label: 'Saturation', value: 'mix-blend-saturation' },
+	{ label: 'Color', value: 'mix-blend-color' },
+	{ label: 'Luminosity', value: 'mix-blend-luminosity' },
+];
