+++ data/bootstrap-classes/classes.js
@@ -1,15 +1,4 @@
 // options.js
-export const bleedCoverOptions = [
-	{
-		label: 'Extend Cover Left',
-		value: 'cover-negative-margin-left',
-	},
-	{
-		label: 'Extend Cover Right',
-		value: 'cover-negative-margin-right',
-	},
-];
-
 export const columnOptions = [
 	{ label: 'Default Column', value: 'col' },
 	{ label: 'Auto Width Column', value: 'col-auto' },
@@ -174,75 +163,6 @@ export const cssGridColumnOptions = [
 	{ label: 'Grid XXL Full Width', value: 'g-col-xxl-12' },
 ];
 
-export const displayOptions = [
-	{ label: 'Hide', value: 'd-none' },
-	{ label: 'Block Display', value: 'd-block' },
-	{ label: 'Inline Display', value: 'd-inline' },
-	{ label: 'Inline Block', value: 'd-inline-block' },
-	{ label: 'Flex Display', value: 'd-flex' },
-	{ label: 'Inline Flex', value: 'd-inline-flex' },
-	{ label: 'Table Display', value: 'd-table' },
-	{ label: 'Table Row', value: 'd-table-row' },
-	{ label: 'Table Cell', value: 'd-table-cell' },
-	{ label: 'Grid Display', value: 'd-grid' },
-	{ label: 'Inline Grid', value: 'd-inline-grid' },
-	{ label: 'Hide on Small', value: 'd-sm-none' },
-	{ label: 'Block on Small', value: 'd-sm-block' },
-	{ label: 'Inline on Small', value: 'd-sm-inline' },
-	{ label: 'Inline Block on Small', value: 'd-sm-inline-block' },
-	{ label: 'Flex on Small', value: 'd-sm-flex' },
-	{ label: 'Inline Flex on Small', value: 'd-sm-inline-flex' },
-	{ label: 'Table on Small', value: 'd-sm-table' },
-	{ label: 'Table Row on Small', value: 'd-sm-table-row' },
-	{ label: 'Table Cell on Small', value: 'd-sm-table-cell' },
-	{ label: 'Grid on Small', value: 'd-sm-grid' },
-	{ label: 'Inline Grid on Small', value: 'd-sm-inline-grid' },
-	{ label: 'Hide on Medium', value: 'd-md-none' },
-	{ label: 'Block on Medium', value: 'd-md-block' },
-	{ label: 'Inline on Medium', value: 'd-md-inline' },
-	{ label: 'Inline Block on Medium', value: 'd-md-inline-block' },
-	{ label: 'Flex on Medium', value: 'd-md-flex' },
-	{ label: 'Inline Flex on Medium', value: 'd-md-inline-flex' },
-	{ label: 'Table on Medium', value: 'd-md-table' },
-	{ label: 'Table Row on Medium', value: 'd-md-table-row' },
-	{ label: 'Table Cell on Medium', value: 'd-md-table-cell' },
-	{ label: 'Grid on Medium', value: 'd-md-grid' },
-	{ label: 'Inline Grid on Medium', value: 'd-md-inline-grid' },
-	{ label: 'Hide on Large', value: 'd-lg-none' },
-	{ label: 'Block on Large', value: 'd-lg-block' },
-	{ label: 'Inline on Large', value: 'd-lg-inline' },
-	{ label: 'Inline Block on Large', value: 'd-lg-inline-block' },
-	{ label: 'Flex on Large', value: 'd-lg-flex' },
-	{ label: 'Inline Flex on Large', value: 'd-lg-inline-flex' },
-	{ label: 'Table on Large', value: 'd-lg-table' },
-	{ label: 'Table Row on Large', value: 'd-lg-table-row' },
-	{ label: 'Table Cell on Large', value: 'd-lg-table-cell' },
-	{ label: 'Grid on Large', value: 'd-lg-grid' },
-	{ label: 'Inline Grid on Large', value: 'd-lg-inline-grid' },
-	{ label: 'Hide on Extra Large', value: 'd-xl-none' },
-	{ label: 'Block on Extra Large', value: 'd-xl-block' },
-	{ label: 'Inline on Extra Large', value: 'd-xl-inline' },
-	{ label: 'Inline Block on Extra Large', value: 'd-xl-inline-block' },
-	{ label: 'Flex on Extra Large', value: 'd-xl-flex' },
-	{ label: 'Inline Flex on Extra Large', value: 'd-xl-inline-flex' },
-	{ label: 'Table on Extra Large', value: 'd-xl-table' },
-	{ label: 'Table Row on Extra Large', value: 'd-xl-table-row' },
-	{ label: 'Table Cell on Extra Large', value: 'd-xl-table-cell' },
-	{ label: 'Grid on Extra Large', value: 'd-xl-grid' },
-	{ label: 'Inline Grid on Extra Large', value: 'd-xl-inline-grid' },
-	{ label: 'Hide on XXL', value: 'd-xxl-none' },
-	{ label: 'Block on XXL', value: 'd-xxl-block' },
-	{ label: 'Inline on XXL', value: 'd-xxl-inline' },
-	{ label: 'Inline Block on XXL', value: 'd-xxl-inline-block' },
-	{ label: 'Flex on XXL', value: 'd-xxl-flex' },
-	{ label: 'Inline Flex on XXL', value: 'd-xxl-inline-flex' },
-	{ label: 'Table on XXL', value: 'd-xxl-table' },
-	{ label: 'Table Row on XXL', value: 'd-xxl-table-row' },
-	{ label: 'Table Cell on XXL', value: 'd-xxl-table-cell' },
-	{ label: 'Grid on XXL', value: 'd-xxl-grid' },
-	{ label: 'Inline Grid on XXL', value: 'd-xxl-inline-grid' },
-];
-
 export const overflowOptions = [
 	{ label: 'Auto Overflow', value: 'overflow-auto' },
 	{ label: 'Hide Overflow', value: 'overflow-hidden' },
@@ -256,392 +176,6 @@ export const overflowOptions = [
 	{ label: 'Vertical Hide Overflow', value: 'overflow-y-hidden' },
 	{ label: 'Vertical Visible Overflow', value: 'overflow-y-visible' },
 	{ label: 'Vertical Scroll Overflow', value: 'overflow-y-scroll' },
-];
-
-export const selfAlignmentOptions = [
-	{ label: 'Align Top', value: 'align-self-start' },
-	{ label: 'Medium Screen Align Top', value: 'align-self-md-start' },
-	{ label: 'Large Screen Align Top', value: 'align-self-lg-start' },
-	{ label: 'Extra Large Align Top', value: 'align-self-xl-start' },
-	{ label: 'XXL Align Top', value: 'align-self-xxl-start' },
-	{ label: 'Align Center', value: 'align-self-center' },
-	{ label: 'Medium Screen Align Center', value: 'align-self-md-center' },
-	{ label: 'Large Screen Align Center', value: 'align-self-lg-center' },
-	{ label: 'Extra Large Align Center', value: 'align-self-xl-center' },
-	{ label: 'XXL Align Center', value: 'align-self-xxl-center' },
-	{ label: 'Align Bottom', value: 'align-self-end' },
-	{ label: 'Medium Screen Align Bottom', value: 'align-self-md-end' },
-	{ label: 'Large Screen Align Bottom', value: 'align-self-lg-end' },
-	{ label: 'Extra Large Align Bottom', value: 'align-self-xl-end' },
-	{ label: 'XXL Align Bottom', value: 'align-self-xxl-end' },
-	{ label: 'Align Baseline', value: 'align-self-baseline' },
-	{ label: 'Medium Screen Align Baseline', value: 'align-self-md-baseline' },
-	{ label: 'Large Screen Align Baseline', value: 'align-self-lg-baseline' },
-	{ label: 'Extra Large Align Baseline', value: 'align-self-xl-baseline' },
-	{ label: 'XXL Align Baseline', value: 'align-self-xxl-baseline' },
-	{ label: 'Stretch Align', value: 'align-self-stretch' },
-	{ label: 'Medium Screen Stretch', value: 'align-self-md-stretch' },
-	{ label: 'Large Screen Stretch', value: 'align-self-lg-stretch' },
-	{ label: 'Extra Large Stretch', value: 'align-self-xl-stretch' },
-	{ label: 'XXL Stretch', value: 'align-self-xxl-stretch' },
-];
-
-export const marginOptions = [
-	{ label: 'Start Margin Auto', value: 'ms-auto' },
-	{ label: 'End Margin Auto', value: 'me-auto' },
-	{ label: 'Small Screen End Margin Auto', value: 'me-sm-auto' },
-	{ label: 'Medium Screen End Margin Auto', value: 'me-md-auto' },
-	{ label: 'Medium Screen Start Margin Auto', value: 'ms-md-auto' },
-	{ label: 'Large Screen End Margin Auto', value: 'me-lg-auto' },
-	{ label: 'Large Screen Start Margin Auto', value: 'ms-lg-auto' },
-	{ label: 'Extra Large End Margin Auto', value: 'me-xl-auto' },
-	{ label: 'Extra Large Start Margin Auto', value: 'ms-xl-auto' },
-	{ label: 'XXL End Margin Auto', value: 'me-xxl-auto' },
-	{ label: 'XXL Start Margin Auto', value: 'ms-xxl-auto' },
-	{ label: 'No Margin', value: 'm-0' },
-	{ label: 'Margin 1', value: 'm-1' },
-	{ label: 'Margin 2', value: 'm-2' },
-	{ label: 'Margin 3', value: 'm-3' },
-	{ label: 'Margin 4', value: 'm-4' },
-	{ label: 'Margin 5', value: 'm-5' },
-	{ label: 'Margin 6', value: 'm-6' },
-	{ label: 'Margin 7', value: 'm-7' },
-	{ label: 'Margin 8', value: 'm-8' },
-	{ label: 'Margin 9', value: 'm-9' },
-	{ label: 'Margin 10', value: 'm-10' },
-	{ label: 'Bottom Margin 0', value: 'mb-0' },
-	{ label: 'Bottom Margin 1', value: 'mb-1' },
-	{ label: 'Bottom Margin 2', value: 'mb-2' },
-	{ label: 'Bottom Margin 3', value: 'mb-3' },
-	{ label: 'Bottom Margin 4', value: 'mb-4' },
-	{ label: 'Bottom Margin 5', value: 'mb-5' },
-	{ label: 'Bottom Margin 6', value: 'mb-6' },
-	{ label: 'Bottom Margin 7', value: 'mb-7' },
-	{ label: 'Bottom Margin 8', value: 'mb-8' },
-	{ label: 'Bottom Margin 9', value: 'mb-9' },
-	{ label: 'Bottom Margin 10', value: 'mb-10' },
-	{ label: 'Top Margin 0', value: 'mt-0' },
-	{ label: 'Top Margin 1', value: 'mt-1' },
-	{ label: 'Top Margin 2', value: 'mt-2' },
-	{ label: 'Top Margin 3', value: 'mt-3' },
-	{ label: 'Top Margin 4', value: 'mt-4' },
-	{ label: 'Top Margin 5', value: 'mt-5' },
-	{ label: 'Top Margin 6', value: 'mt-6' },
-	{ label: 'Top Margin 7', value: 'mt-7' },
-	{ label: 'Top Margin 8', value: 'mt-8' },
-	{ label: 'Top Margin 9', value: 'mt-9' },
-	{ label: 'Top Margin 10', value: 'mt-10' },
-	{ label: 'Negative Start Margin 0', value: 'ms-n0' },
-	{ label: 'Negative Start Margin 1', value: 'ms-n1' },
-	{ label: 'Negative Start Margin 2', value: 'ms-n2' },
-	{ label: 'Negative Start Margin 3', value: 'ms-n3' },
-	{ label: 'Negative Start Margin 4', value: 'ms-n4' },
-	{ label: 'Negative Start Margin 5', value: 'ms-n5' },
-	{ label: 'Negative Start Margin 6', value: 'ms-n6' },
-	{ label: 'Negative Start Margin 7', value: 'ms-n7' },
-	{ label: 'Negative Start Margin 8', value: 'ms-n8' },
-	{ label: 'Negative Start Margin 9', value: 'ms-n9' },
-	{ label: 'Negative Start Margin 10', value: 'ms-n10' },
-	{ label: 'Negative End Margin 0', value: 'me-n0' },
-	{ label: 'Negative End Margin 1', value: 'me-n1' },
-	{ label: 'Negative End Margin 2', value: 'me-n2' },
-	{ label: 'Negative End Margin 3', value: 'me-n3' },
-	{ label: 'Negative End Margin 4', value: 'me-n4' },
-	{ label: 'Negative End Margin 5', value: 'me-n5' },
-	{ label: 'Negative End Margin 6', value: 'me-n6' },
-	{ label: 'Negative End Margin 7', value: 'me-n7' },
-	{ label: 'Negative End Margin 8', value: 'me-n8' },
-	{ label: 'Negative End Margin 9', value: 'me-n9' },
-	{ label: 'Negative End Margin 10', value: 'me-n10' },
-	{ label: 'Medium Negative Start Margin 0', value: 'ms-md-n0' },
-	{ label: 'Medium Negative Start Margin 1', value: 'ms-md-n1' },
-	{ label: 'Medium Negative Start Margin 2', value: 'ms-md-n2' },
-	{ label: 'Medium Negative Start Margin 3', value: 'ms-md-n3' },
-	{ label: 'Medium Negative Start Margin 4', value: 'ms-md-n4' },
-	{ label: 'Medium Negative Start Margin 5', value: 'ms-md-n5' },
-	{ label: 'Medium Negative Start Margin 6', value: 'ms-md-n6' },
-	{ label: 'Medium Negative Start Margin 7', value: 'ms-md-n7' },
-	{ label: 'Medium Negative Start Margin 8', value: 'ms-md-n8' },
-	{ label: 'Medium Negative Start Margin 9', value: 'ms-md-n9' },
-	{ label: 'Medium Negative Start Margin 10', value: 'ms-md-n10' },
-	{ label: 'Large Negative Start Margin 0', value: 'ms-lg-n0' },
-	{ label: 'Large Negative Start Margin 1', value: 'ms-lg-n1' },
-	{ label: 'Large Negative Start Margin 2', value: 'ms-lg-n2' },
-	{ label: 'Large Negative Start Margin 3', value: 'ms-lg-n3' },
-	{ label: 'Large Negative Start Margin 4', value: 'ms-lg-n4' },
-	{ label: 'Large Negative Start Margin 5', value: 'ms-lg-n5' },
-	{ label: 'Large Negative Start Margin 6', value: 'ms-lg-n6' },
-	{ label: 'Large Negative Start Margin 7', value: 'ms-lg-n7' },
-	{ label: 'Large Negative Start Margin 8', value: 'ms-lg-n8' },
-	{ label: 'Large Negative Start Margin 9', value: 'ms-lg-n9' },
-	{ label: 'Large Negative Start Margin 10', value: 'ms-lg-n10' },
-	{ label: 'Extra Large Negative Start Margin 0', value: 'ms-xl-n0' },
-	{ label: 'Extra Large Negative Start Margin 1', value: 'ms-xl-n1' },
-	{ label: 'Extra Large Negative Start Margin 2', value: 'ms-xl-n2' },
-	{ label: 'Extra Large Negative Start Margin 3', value: 'ms-xl-n3' },
-	{ label: 'Extra Large Negative Start Margin 4', value: 'ms-xl-n4' },
-	{ label: 'Extra Large Negative Start Margin 5', value: 'ms-xl-n5' },
-	{ label: 'Extra Large Negative Start Margin 6', value: 'ms-xl-n6' },
-	{ label: 'Extra Large Negative Start Margin 7', value: 'ms-xl-n7' },
-	{ label: 'Extra Large Negative Start Margin 8', value: 'ms-xl-n8' },
-	{ label: 'Extra Large Negative Start Margin 9', value: 'ms-xl-n9' },
-	{ label: 'Extra Large Negative Start Margin 10', value: 'ms-xl-n10' },
-	{ label: 'XXL Negative Start Margin 0', value: 'ms-xxl-n0' },
-	{ label: 'XXL Negative Start Margin 1', value: 'ms-xxl-n1' },
-	{ label: 'XXL Negative Start Margin 2', value: 'ms-xxl-n2' },
-	{ label: 'XXL Negative Start Margin 3', value: 'ms-xxl-n3' },
-	{ label: 'XXL Negative Start Margin 4', value: 'ms-xxl-n4' },
-	{ label: 'XXL Negative Start Margin 5', value: 'ms-xxl-n5' },
-	{ label: 'XXL Negative Start Margin 6', value: 'ms-xxl-n6' },
-	{ label: 'XXL Negative Start Margin 7', value: 'ms-xxl-n7' },
-	{ label: 'XXL Negative Start Margin 8', value: 'ms-xxl-n8' },
-	{ label: 'XXL Negative Start Margin 9', value: 'ms-xxl-n9' },
-	{ label: 'XXL Negative Start Margin 10', value: 'ms-xxl-n10' },
-	{ label: 'Medium Negative End Margin 0', value: 'me-md-n0' },
-	{ label: 'Medium Negative End Margin 1', value: 'me-md-n1' },
-	{ label: 'Medium Negative End Margin 2', value: 'me-md-n2' },
-	{ label: 'Medium Negative End Margin 3', value: 'me-md-n3' },
-	{ label: 'Medium Negative End Margin 4', value: 'me-md-n4' },
-	{ label: 'Medium Negative End Margin 5', value: 'me-md-n5' },
-	{ label: 'Medium Negative End Margin 6', value: 'me-md-n6' },
-	{ label: 'Medium Negative End Margin 7', value: 'me-md-n7' },
-	{ label: 'Medium Negative End Margin 8', value: 'me-md-n8' },
-	{ label: 'Medium Negative End Margin 9', value: 'me-md-n9' },
-	{ label: 'Medium Negative End Margin 10', value: 'me-md-n10' },
-	{ label: 'Large Negative End Margin 0', value: 'me-lg-n0' },
-	{ label: 'Large Negative End Margin 1', value: 'me-lg-n1' },
-	{ label: 'Large Negative End Margin 2', value: 'me-lg-n2' },
-	{ label: 'Large Negative End Margin 3', value: 'me-lg-n3' },
-	{ label: 'Large Negative End Margin 4', value: 'me-lg-n4' },
-	{ label: 'Large Negative End Margin 5', value: 'me-lg-n5' },
-	{ label: 'Large Negative End Margin 6', value: 'me-lg-n6' },
-	{ label: 'Large Negative End Margin 7', value: 'me-lg-n7' },
-	{ label: 'Large Negative End Margin 8', value: 'me-lg-n8' },
-	{ label: 'Large Negative End Margin 9', value: 'me-lg-n9' },
-	{ label: 'Large Negative End Margin 10', value: 'me-lg-n10' },
-	{ label: 'Extra Large Negative End Margin 0', value: 'me-xl-n0' },
-	{ label: 'Extra Large Negative End Margin 1', value: 'me-xl-n1' },
-	{ label: 'Extra Large Negative End Margin 2', value: 'me-xl-n2' },
-	{ label: 'Extra Large Negative End Margin 3', value: 'me-xl-n3' },
-	{ label: 'Extra Large Negative End Margin 4', value: 'me-xl-n4' },
-	{ label: 'Extra Large Negative End Margin 5', value: 'me-xl-n5' },
-	{ label: 'Extra Large Negative End Margin 6', value: 'me-xl-n6' },
-	{ label: 'Extra Large Negative End Margin 7', value: 'me-xl-n7' },
-	{ label: 'Extra Large Negative End Margin 8', value: 'me-xl-n8' },
-	{ label: 'Extra Large Negative End Margin 9', value: 'me-xl-n9' },
-	{ label: 'Extra Large Negative End Margin 10', value: 'me-xl-n10' },
-	{ label: 'XXL Negative End Margin 0', value: 'me-xxl-n0' },
-	{ label: 'XXL Negative End Margin 1', value: 'me-xxl-n1' },
-	{ label: 'XXL Negative End Margin 2', value: 'me-xxl-n2' },
-	{ label: 'XXL Negative End Margin 3', value: 'me-xxl-n3' },
-	{ label: 'XXL Negative End Margin 4', value: 'me-xxl-n4' },
-	{ label: 'XXL Negative End Margin 5', value: 'me-xxl-n5' },
-	{ label: 'XXL Negative End Margin 6', value: 'me-xxl-n6' },
-	{ label: 'XXL Negative End Margin 7', value: 'me-xxl-n7' },
-	{ label: 'XXL Negative End Margin 8', value: 'me-xxl-n8' },
-	{ label: 'XXL Negative End Margin 9', value: 'me-xxl-n9' },
-	{ label: 'XXL Negative End Margin 10', value: 'me-xxl-n10' },
-];
-
-export const paddingOptions = [
-	{ label: 'No Padding', value: 'p-0' },
-	{ label: 'Padding 1', value: 'p-1' },
-	{ label: 'Padding 2', value: 'p-2' },
-	{ label: 'Padding 3', value: 'p-3' },
-	{ label: 'Padding 4', value: 'p-4' },
-	{ label: 'Padding 5', value: 'p-5' },
-	{ label: 'Padding 6', value: 'p-6' },
-	{ label: 'Padding 7', value: 'p-7' },
-	{ label: 'Padding 8', value: 'p-8' },
-	{ label: 'Padding 9', value: 'p-9' },
-	{ label: 'Padding 10', value: 'p-10' },
-	{ label: 'Small Screen No Padding', value: 'p-sm-0' },
-	{ label: 'Small Screen Padding 1', value: 'p-sm-1' },
-	{ label: 'Small Screen Padding 2', value: 'p-sm-2' },
-	{ label: 'Small Screen Padding 3', value: 'p-sm-3' },
-	{ label: 'Small Screen Padding 4', value: 'p-sm-4' },
-	{ label: 'Small Screen Padding 5', value: 'p-sm-5' },
-	{ label: 'Small Screen Padding 6', value: 'p-sm-6' },
-	{ label: 'Small Screen Padding 7', value: 'p-sm-7' },
-	{ label: 'Small Screen Padding 8', value: 'p-sm-8' },
-	{ label: 'Small Screen Padding 9', value: 'p-sm-9' },
-	{ label: 'Small Screen Padding 10', value: 'p-sm-10' },
-	{ label: 'Medium Screen No Padding', value: 'p-md-0' },
-	{ label: 'Medium Screen Padding 1', value: 'p-md-1' },
-	{ label: 'Medium Screen Padding 2', value: 'p-md-2' },
-	{ label: 'Medium Screen Padding 3', value: 'p-md-3' },
-	{ label: 'Medium Screen Padding 4', value: 'p-md-4' },
-	{ label: 'Medium Screen Padding 5', value: 'p-md-5' },
-	{ label: 'Medium Screen Padding 6', value: 'p-md-6' },
-	{ label: 'Medium Screen Padding 7', value: 'p-md-7' },
-	{ label: 'Medium Screen Padding 8', value: 'p-md-8' },
-	{ label: 'Medium Screen Padding 9', value: 'p-md-9' },
-	{ label: 'Medium Screen Padding 10', value: 'p-md-10' },
-	{ label: 'Large Screen No Padding', value: 'p-lg-0' },
-	{ label: 'Large Screen Padding 1', value: 'p-lg-1' },
-	{ label: 'Large Screen Padding 2', value: 'p-lg-2' },
-	{ label: 'Large Screen Padding 3', value: 'p-lg-3' },
-	{ label: 'Large Screen Padding 4', value: 'p-lg-4' },
-	{ label: 'Large Screen Padding 5', value: 'p-lg-5' },
-	{ label: 'Large Screen Padding 6', value: 'p-lg-6' },
-	{ label: 'Large Screen Padding 7', value: 'p-lg-7' },
-	{ label: 'Large Screen Padding 8', value: 'p-lg-8' },
-	{ label: 'Large Screen Padding 9', value: 'p-lg-9' },
-	{ label: 'Large Screen Padding 10', value: 'p-lg-10' },
-	{ label: 'Extra Large No Padding', value: 'p-xl-0' },
-	{ label: 'Extra Large Padding 1', value: 'p-xl-1' },
-	{ label: 'Extra Large Padding 2', value: 'p-xl-2' },
-	{ label: 'Extra Large Padding 3', value: 'p-xl-3' },
-	{ label: 'Extra Large Padding 4', value: 'p-xl-4' },
-	{ label: 'Extra Large Padding 5', value: 'p-xl-5' },
-	{ label: 'Extra Large Padding 6', value: 'p-xl-6' },
-	{ label: 'Extra Large Padding 7', value: 'p-xl-7' },
-	{ label: 'Extra Large Padding 8', value: 'p-xl-8' },
-	{ label: 'Extra Large Padding 9', value: 'p-xl-9' },
-	{ label: 'Extra Large Padding 10', value: 'p-xl-10' },
-	{ label: 'XXL No Padding', value: 'p-xxl-0' },
-	{ label: 'XXL Padding 1', value: 'p-xxl-1' },
-	{ label: 'XXL Padding 2', value: 'p-xxl-2' },
-	{ label: 'XXL Padding 3', value: 'p-xxl-3' },
-	{ label: 'XXL Padding 4', value: 'p-xxl-4' },
-	{ label: 'XXL Padding 5', value: 'p-xxl-5' },
-	{ label: 'XXL Padding 6', value: 'p-xxl-6' },
-	{ label: 'XXL Padding 7', value: 'p-xxl-7' },
-	{ label: 'XXL Padding 8', value: 'p-xxl-8' },
-	{ label: 'XXL Padding 9', value: 'p-xxl-9' },
-	{ label: 'XXL Padding 10', value: 'p-xxl-10' },
-	{ label: 'Top Padding 0', value: 'pt-0' },
-	{ label: 'Top Padding 1', value: 'pt-1' },
-	{ label: 'Top Padding 2', value: 'pt-2' },
-	{ label: 'Top Padding 3', value: 'pt-3' },
-	{ label: 'Top Padding 4', value: 'pt-4' },
-	{ label: 'Top Padding 5', value: 'pt-5' },
-	{ label: 'Top Padding 6', value: 'pt-6' },
-	{ label: 'Top Padding 7', value: 'pt-7' },
-	{ label: 'Top Padding 8', value: 'pt-8' },
-	{ label: 'Top Padding 9', value: 'pt-9' },
-	{ label: 'Top Padding 10', value: 'pt-10' },
-	{ label: 'Small Screen Top Padding 0', value: 'pt-sm-0' },
-	{ label: 'Small Screen Top Padding 1', value: 'pt-sm-1' },
-	{ label: 'Small Screen Top Padding 2', value: 'pt-sm-2' },
-	{ label: 'Small Screen Top Padding 3', value: 'pt-sm-3' },
-	{ label: 'Small Screen Top Padding 4', value: 'pt-sm-4' },
-	{ label: 'Small Screen Top Padding 5', value: 'pt-sm-5' },
-	{ label: 'Small Screen Top Padding 6', value: 'pt-sm-6' },
-	{ label: 'Small Screen Top Padding 7', value: 'pt-sm-7' },
-	{ label: 'Small Screen Top Padding 8', value: 'pt-sm-8' },
-	{ label: 'Small Screen Top Padding 9', value: 'pt-sm-9' },
-	{ label: 'Small Screen Top Padding 10', value: 'pt-sm-10' },
-	{ label: 'Medium Screen Top Padding 0', value: 'pt-md-0' },
-	{ label: 'Medium Screen Top Padding 1', value: 'pt-md-1' },
-	{ label: 'Medium Screen Top Padding 2', value: 'pt-md-2' },
-	{ label: 'Medium Screen Top Padding 3', value: 'pt-md-3' },
-	{ label: 'Medium Screen Top Padding 4', value: 'pt-md-4' },
-	{ label: 'Medium Screen Top Padding 5', value: 'pt-md-5' },
-	{ label: 'Medium Screen Top Padding 6', value: 'pt-md-6' },
-	{ label: 'Medium Screen Top Padding 7', value: 'pt-md-7' },
-	{ label: 'Medium Screen Top Padding 8', value: 'pt-md-8' },
-	{ label: 'Medium Screen Top Padding 9', value: 'pt-md-9' },
-	{ label: 'Medium Screen Top Padding 10', value: 'pt-md-10' },
-	{ label: 'Large Screen Top Padding 0', value: 'pt-lg-0' },
-	{ label: 'Large Screen Top Padding 1', value: 'pt-lg-1' },
-	{ label: 'Large Screen Top Padding 2', value: 'pt-lg-2' },
-	{ label: 'Large Screen Top Padding 3', value: 'pt-lg-3' },
-	{ label: 'Large Screen Top Padding 4', value: 'pt-lg-4' },
-	{ label: 'Large Screen Top Padding 5', value: 'pt-lg-5' },
-	{ label: 'Large Screen Top Padding 6', value: 'pt-lg-6' },
-	{ label: 'Large Screen Top Padding 7', value: 'pt-lg-7' },
-	{ label: 'Large Screen Top Padding 8', value: 'pt-lg-8' },
-	{ label: 'Large Screen Top Padding 9', value: 'pt-lg-9' },
-	{ label: 'Large Screen Top Padding 10', value: 'pt-lg-10' },
-	{ label: 'Extra Large Top Padding 0', value: 'pt-xl-0' },
-	{ label: 'Extra Large Top Padding 1', value: 'pt-xl-1' },
-	{ label: 'Extra Large Top Padding 2', value: 'pt-xl-2' },
-	{ label: 'Extra Large Top Padding 3', value: 'pt-xl-3' },
-	{ label: 'Extra Large Top Padding 4', value: 'pt-xl-4' },
-	{ label: 'Extra Large Top Padding 5', value: 'pt-xl-5' },
-	{ label: 'Extra Large Top Padding 6', value: 'pt-xl-6' },
-	{ label: 'Extra Large Top Padding 7', value: 'pt-xl-7' },
-	{ label: 'Extra Large Top Padding 8', value: 'pt-xl-8' },
-	{ label: 'Extra Large Top Padding 9', value: 'pt-xl-9' },
-	{ label: 'Extra Large Top Padding 10', value: 'pt-xl-10' },
-	{ label: 'XXL Top Padding 0', value: 'pt-xxl-0' },
-	{ label: 'XXL Top Padding 1', value: 'pt-xxl-1' },
-	{ label: 'XXL Top Padding 2', value: 'pt-xxl-2' },
-	{ label: 'XXL Top Padding 3', value: 'pt-xxl-3' },
-	{ label: 'XXL Top Padding 4', value: 'pt-xxl-4' },
-	{ label: 'XXL Top Padding 5', value: 'pt-xxl-5' },
-	{ label: 'XXL Top Padding 6', value: 'pt-xxl-6' },
-	{ label: 'XXL Top Padding 7', value: 'pt-xxl-7' },
-	{ label: 'XXL Top Padding 8', value: 'pt-xxl-8' },
-	{ label: 'XXL Top Padding 9', value: 'pt-xxl-9' },
-	{ label: 'XXL Top Padding 10', value: 'pt-xxl-10' },
-	{ label: 'Bottom Padding 0', value: 'pb-0' },
-	{ label: 'Bottom Padding 1', value: 'pb-1' },
-	{ label: 'Bottom Padding 2', value: 'pb-2' },
-	{ label: 'Bottom Padding 3', value: 'pb-3' },
-	{ label: 'Bottom Padding 4', value: 'pb-4' },
-	{ label: 'Bottom Padding 5', value: 'pb-5' },
-	{ label: 'Bottom Padding 6', value: 'pb-6' },
-	{ label: 'Bottom Padding 7', value: 'pb-7' },
-	{ label: 'Bottom Padding 8', value: 'pb-8' },
-	{ label: 'Bottom Padding 9', value: 'pb-9' },
-	{ label: 'Bottom Padding 10', value: 'pb-10' },
-	{ label: 'Small Screen Bottom Padding 0', value: 'pb-sm-0' },
-	{ label: 'Small Screen Bottom Padding 1', value: 'pb-sm-1' },
-	{ label: 'Small Screen Bottom Padding 2', value: 'pb-sm-2' },
-	{ label: 'Small Screen Bottom Padding 3', value: 'pb-sm-3' },
-	{ label: 'Small Screen Bottom Padding 4', value: 'pb-sm-4' },
-	{ label: 'Small Screen Bottom Padding 5', value: 'pb-sm-5' },
-	{ label: 'Small Screen Bottom Padding 6', value: 'pb-sm-6' },
-	{ label: 'Small Screen Bottom Padding 7', value: 'pb-sm-7' },
-	{ label: 'Small Screen Bottom Padding 8', value: 'pb-sm-8' },
-	{ label: 'Small Screen Bottom Padding 9', value: 'pb-sm-9' },
-	{ label: 'Small Screen Bottom Padding 10', value: 'pb-sm-10' },
-	{ label: 'Medium Screen Bottom Padding 0', value: 'pb-md-0' },
-	{ label: 'Medium Screen Bottom Padding 1', value: 'pb-md-1' },
-	{ label: 'Medium Screen Bottom Padding 2', value: 'pb-md-2' },
-	{ label: 'Medium Screen Bottom Padding 3', value: 'pb-md-3' },
-	{ label: 'Medium Screen Bottom Padding 4', value: 'pb-md-4' },
-	{ label: 'Medium Screen Bottom Padding 5', value: 'pb-md-5' },
-	{ label: 'Medium Screen Bottom Padding 6', value: 'pb-md-6' },
-	{ label: 'Medium Screen Bottom Padding 7', value: 'pb-md-7' },
-	{ label: 'Medium Screen Bottom Padding 8', value: 'pb-md-8' },
-	{ label: 'Medium Screen Bottom Padding 9', value: 'pb-md-9' },
-	{ label: 'Medium Screen Bottom Padding 10', value: 'pb-md-10' },
-	{ label: 'Large Screen Bottom Padding 0', value: 'pb-lg-0' },
-	{ label: 'Large Screen Bottom Padding 1', value: 'pb-lg-1' },
-	{ label: 'Large Screen Bottom Padding 2', value: 'pb-lg-2' },
-	{ label: 'Large Screen Bottom Padding 3', value: 'pb-lg-3' },
-	{ label: 'Large Screen Bottom Padding 4', value: 'pb-lg-4' },
-	{ label: 'Large Screen Bottom Padding 5', value: 'pb-lg-5' },
-	{ label: 'Large Screen Bottom Padding 6', value: 'pb-lg-6' },
-	{ label: 'Large Screen Bottom Padding 7', value: 'pb-lg-7' },
-	{ label: 'Large Screen Bottom Padding 8', value: 'pb-lg-8' },
-	{ label: 'Large Screen Bottom Padding 9', value: 'pb-lg-9' },
-	{ label: 'Large Screen Bottom Padding 10', value: 'pb-lg-10' },
-	{ label: 'Extra Large Bottom Padding 0', value: 'pb-xl-0' },
-	{ label: 'Extra Large Bottom Padding 1', value: 'pb-xl-1' },
-	{ label: 'Extra Large Bottom Padding 2', value: 'pb-xl-2' },
-	{ label: 'Extra Large Bottom Padding 3', value: 'pb-xl-3' },
-	{ label: 'Extra Large Bottom Padding 4', value: 'pb-xl-4' },
-	{ label: 'Extra Large Bottom Padding 5', value: 'pb-xl-5' },
-	{ label: 'Extra Large Bottom Padding 6', value: 'pb-xl-6' },
-	{ label: 'Extra Large Bottom Padding 7', value: 'pb-xl-7' },
-	{ label: 'Extra Large Bottom Padding 8', value: 'pb-xl-8' },
-	{ label: 'Extra Large Bottom Padding 9', value: 'pb-xl-9' },
-	{ label: 'Extra Large Bottom Padding 10', value: 'pb-xl-10' },
-	{ label: 'XXL Bottom Padding 0', value: 'pb-xxl-0' },
-	{ label: 'XXL Bottom Padding 1', value: 'pb-xxl-1' },
-	{ label: 'XXL Bottom Padding 2', value: 'pb-xxl-2' },
-	{ label: 'XXL Bottom Padding 3', value: 'pb-xxl-3' },
-	{ label: 'XXL Bottom Padding 4', value: 'pb-xxl-4' },
-	{ label: 'XXL Bottom Padding 5', value: 'pb-xxl-5' },
-	{ label: 'XXL Bottom Padding 6', value: 'pb-xxl-6' },
-	{ label: 'XXL Bottom Padding 7', value: 'pb-xxl-7' },
-	{ label: 'XXL Bottom Padding 8', value: 'pb-xxl-8' },
-	{ label: 'XXL Bottom Padding 9', value: 'pb-xxl-9' },
-	{ label: 'XXL Bottom Padding 10', value: 'pb-xxl-10' },
 ];
 
 export const rowOptions = [
@@ -671,36 +205,6 @@ export const rowOptions = [
 	{ label: 'Large Screen 6 Columns', value: 'row-cols-lg-6' },
 	{ label: 'Extra Large 6 Columns', value: 'row-cols-xl-6' },
 	{ label: 'XXL 6 Columns', value: 'row-cols-xxl-6' },
-];
-
-export const gapOptions = [
-	{ label: 'No Gap', value: 'gap-0' },
-	{ label: 'Gap 1', value: 'gap-1' },
-	{ label: 'Gap 2', value: 'gap-2' },
-	{ label: 'Gap 3', value: 'gap-3' },
-	{ label: 'Gap 4', value: 'gap-4' },
-	{ label: 'Gap 5', value: 'gap-5' },
-	{ label: 'Small Screen No Gap', value: 'gap-sm-0' },
-	{ label: 'Small Screen Gap 1', value: 'gap-sm-1' },
-	{ label: 'Small Screen Gap 2', value: 'gap-sm-2' },
-	{ label: 'Small Screen Gap 3', value: 'gap-sm-3' },
-	{ label: 'Small Screen Gap 4', value: 'gap-sm-4' },
-	{ label: 'Small Screen Gap 5', value: 'gap-sm-5' },
-	{ label: 'Medium Screen No Gap', value: 'gap-md-0' },
-	{ label: 'Medium Screen Gap 1', value: 'gap-md-1' },
-	{ label: 'Medium Screen Gap 2', value: 'gap-md-2' },
-	{ label: 'Medium Screen Gap 3', value: 'gap-md-3' },
-	{ label: 'Medium Screen Gap 4', value: 'gap-md-4' },
-	{ label: 'Medium Screen Gap 5', value: 'gap-md-5' },
-	{ label: 'Large Screen No Gap', value: 'gap-lg-0' },
-	{ label: 'Large Screen Gap 1', value: 'gap-lg-1' },
-	{ label: 'Large Screen Gap 2', value: 'gap-lg-2' },
-	{ label: 'Large Screen Gap 3', value: 'gap-lg-3' },
-	{ label: 'Large Screen Gap 4', value: 'gap-lg-4' },
-	{ label: 'Large Screen Gap 5', value: 'gap-lg-5' },
-	{ label: 'Extra Large No Gap', value: 'gap-xl-0' },
-	{ label: 'Extra Large Gap 1', value: 'gap-xl-1' },
-	{ label: 'Extra Large Gap 2', value: 'gap-xl-2' },
 ];
 
 export const columnGapOptions = [
@@ -761,123 +265,6 @@ export const rowGapOptions = [
 	{ label: 'Extra Large No Row Gap', value: 'row-gap-xl-0' },
 	{ label: 'Extra Large Row Gap 1', value: 'row-gap-xl-1' },
 	{ label: 'Extra Large Row Gap 2', value: 'row-gap-xl-2' },
-];
-
-export const justifyContentOptions = [
-	{ label: 'Justify Start', value: 'justify-content-start' },
-	{ label: 'Medium Screen Justify Start', value: 'justify-content-md-start' },
-	{ label: 'Large Screen Justify Start', value: 'justify-content-lg-start' },
-	{ label: 'Extra Large Justify Start', value: 'justify-content-xl-start' },
-	{ label: 'XXL Justify Start', value: 'justify-content-xxl-start' },
-	{ label: 'Justify End', value: 'justify-content-end' },
-	{ label: 'Medium Screen Justify End', value: 'justify-content-md-end' },
-	{ label: 'Large Screen Justify End', value: 'justify-content-lg-end' },
-	{ label: 'Extra Large Justify End', value: 'justify-content-xl-end' },
-	{ label: 'XXL Justify End', value: 'justify-content-xxl-end' },
-	{ label: 'Justify Center', value: 'justify-content-center' },
-	{
-		label: 'Medium Screen Justify Center',
-		value: 'justify-content-md-center',
-	},
-	{
-		label: 'Large Screen Justify Center',
-		value: 'justify-content-lg-center',
-	},
-	{ label: 'Extra Large Justify Center', value: 'justify-content-xl-center' },
-	{ label: 'XXL Justify Center', value: 'justify-content-xxl-center' },
-	{ label: 'Justify Between', value: 'justify-content-between' },
-	{
-		label: 'Medium Screen Justify Between',
-		value: 'justify-content-md-between',
-	},
-	{
-		label: 'Large Screen Justify Between',
-		value: 'justify-content-lg-between',
-	},
-	{
-		label: 'Extra Large Justify Between',
-		value: 'justify-content-xl-between',
-	},
-	{ label: 'XXL Justify Between', value: 'justify-content-xxl-between' },
-];
-
-export const alignItemsOptions = [
-	{ label: 'Align Items Start', value: 'align-items-start' },
-	{ label: 'Medium Screen Align Start', value: 'align-items-md-start' },
-	{ label: 'Large Screen Align Start', value: 'align-items-lg-start' },
-	{ label: 'Extra Large Align Start', value: 'align-items-xl-start' },
-	{ label: 'XXL Align Start', value: 'align-items-xxl-start' },
-	{ label: 'Align Items Center', value: 'align-items-center' },
-	{ label: 'Medium Screen Align Center', value: 'align-items-md-center' },
-	{ label: 'Large Screen Align Center', value: 'align-items-lg-center' },
-	{ label: 'Extra Large Align Center', value: 'align-items-xl-center' },
-	{ label: 'XXL Align Center', value: 'align-items-xxl-center' },
-	{ label: 'Align Items End', value: 'align-items-end' },
-	{ label: 'Medium Screen Align End', value: 'align-items-md-end' },
-	{ label: 'Large Screen Align End', value: 'align-items-lg-end' },
-	{ label: 'Extra Large Align End', value: 'align-items-xl-end' },
-	{ label: 'XXL Align End', value: 'align-items-xxl-end' },
-	{ label: 'Align Items Baseline', value: 'align-items-baseline' },
-	{ label: 'Medium Screen Align Baseline', value: 'align-items-md-baseline' },
-	{ label: 'Large Screen Align Baseline', value: 'align-items-lg-baseline' },
-	{ label: 'Extra Large Align Baseline', value: 'align-items-xl-baseline' },
-	{ label: 'XXL Align Baseline', value: 'align-items-xxl-baseline' },
-	{ label: 'Align Items Stretch', value: 'align-items-stretch' },
-	{ label: 'Medium Screen Align Stretch', value: 'align-items-md-stretch' },
-	{ label: 'Large Screen Align Stretch', value: 'align-items-lg-stretch' },
-	{ label: 'Extra Large Align Stretch', value: 'align-items-xl-stretch' },
-	{ label: 'XXL Align Stretch', value: 'align-items-xxl-stretch' },
-];
-
-export const orderOptions = [
-	{ label: 'Order 0', value: 'order-0' },
-	{ label: 'Order 1', value: 'order-1' },
-	{ label: 'Order 2', value: 'order-2' },
-	{ label: 'Order 3', value: 'order-3' },
-	{ label: 'Order 4', value: 'order-4' },
-	{ label: 'Order 5', value: 'order-5' },
-	{ label: 'Small Screen Order 0', value: 'order-sm-0' },
-	{ label: 'Small Screen Order 1', value: 'order-sm-1' },
-	{ label: 'Small Screen Order 2', value: 'order-sm-2' },
-	{ label: 'Small Screen Order 3', value: 'order-sm-3' },
-	{ label: 'Small Screen Order 4', value: 'order-sm-4' },
-	{ label: 'Small Screen Order 5', value: 'order-sm-5' },
-	{ label: 'Medium Screen Order 0', value: 'order-md-0' },
-	{ label: 'Medium Screen Order 1', value: 'order-md-1' },
-	{ label: 'Medium Screen Order 2', value: 'order-md-2' },
-	{ label: 'Medium Screen Order 3', value: 'order-md-3' },
-	{ label: 'Medium Screen Order 4', value: 'order-md-4' },
-	{ label: 'Medium Screen Order 5', value: 'order-md-5' },
-	{ label: 'Large Screen Order 0', value: 'order-lg-0' },
-	{ label: 'Large Screen Order 1', value: 'order-lg-1' },
-	{ label: 'Large Screen Order 2', value: 'order-lg-2' },
-	{ label: 'Large Screen Order 3', value: 'order-lg-3' },
-	{ label: 'Large Screen Order 4', value: 'order-lg-4' },
-	{ label: 'Large Screen Order 5', value: 'order-lg-5' },
-	{ label: 'Extra Large Order 0', value: 'order-xl-0' },
-	{ label: 'Extra Large Order 1', value: 'order-xl-1' },
-	{ label: 'Extra Large Order 2', value: 'order-xl-2' },
-	{ label: 'Extra Large Order 3', value: 'order-xl-3' },
-	{ label: 'Extra Large Order 4', value: 'order-xl-4' },
-	{ label: 'Extra Large Order 5', value: 'order-xl-5' },
-	{ label: 'XXL Order 0', value: 'order-xxl-0' },
-	{ label: 'XXL Order 1', value: 'order-xxl-1' },
-	{ label: 'XXL Order 2', value: 'order-xxl-2' },
-	{ label: 'XXL Order 3', value: 'order-xxl-3' },
-	{ label: 'XXL Order 4', value: 'order-xxl-4' },
-	{ label: 'XXL Order 5', value: 'order-xxl-5' },
-	{ label: 'Order First', value: 'order-first' },
-	{ label: 'Small Screen Order First', value: 'order-sm-first' },
-	{ label: 'Medium Screen Order First', value: 'order-md-first' },
-	{ label: 'Large Screen Order First', value: 'order-lg-first' },
-	{ label: 'Extra Large Order First', value: 'order-xl-first' },
-	{ label: 'XXL Order First', value: 'order-xxl-first' },
-	{ label: 'Order Last', value: 'order-last' },
-	{ label: 'Small Screen Order Last', value: 'order-sm-last' },
-	{ label: 'Medium Screen Order Last', value: 'order-md-last' },
-	{ label: 'Large Screen Order Last', value: 'order-lg-last' },
-	{ label: 'Extra Large Order Last', value: 'order-xl-last' },
-	{ label: 'XXL Order Last', value: 'order-xxl-last' },
 ];
 
 export const columnOffsetOptions = [
@@ -999,36 +386,6 @@ export const alertOptions = [
 	{ label: 'Dark Alert', value: 'alert-dark' },
 ];
 
-export const positionOptions = [
-	{ label: 'Static Position', value: 'position-static' },
-	{ label: 'Relative Position', value: 'position-relative' },
-	{ label: 'Absolute Position', value: 'position-absolute' },
-	{ label: 'Fixed Position', value: 'position-fixed' },
-	{ label: 'Sticky Position', value: 'position-sticky' },
-	{ label: 'Top 0%', value: 'top-0' },
-	{ label: 'Top 50%', value: 'top-50' },
-	{ label: 'Top 100%', value: 'top-100' },
-	{ label: 'Bottom 0%', value: 'bottom-0' },
-	{ label: 'Bottom 50%', value: 'bottom-50' },
-	{ label: 'Bottom 100%', value: 'bottom-100' },
-	{ label: 'Start 0%', value: 'start-0' },
-	{ label: 'Start 50%', value: 'start-50' },
-	{ label: 'Start 100%', value: 'start-100' },
-	{ label: 'End 0%', value: 'end-0' },
-	{ label: 'End 50%', value: 'end-50' },
-	{ label: 'End 100%', value: 'end-100' },
-	{ label: 'Translate Middle X', value: 'translate-middle-x' },
-	{ label: 'Translate Middle Y', value: 'translate-middle-y' },
-];
-
-export const zindexOptions = [
-	{ label: 'Z-Index 0', value: 'z-0' },
-	{ label: 'Z-Index 1', value: 'z-1' },
-	{ label: 'Z-Index 2', value: 'z-2' },
-	{ label: 'Z-Index 3', value: 'z-3' },
-	{ label: 'Z-Index -1', value: 'z-n1' },
-];
-
 export const borderOptions = [
 	{ label: 'Select a Border', value: '' },
 	{ label: 'No Border', value: 'border-0' },
@@ -1114,25 +471,6 @@ export const borderRadiusOptions = [
 	{ label: 'Large Start Rounding', value: 'rounded-start-3' },
 	{ label: 'Extra Large Start Rounding', value: 'rounded-start-4' },
 	{ label: 'Maximum Start Rounding', value: 'rounded-start-5' },
-];
-
-export const blendModeOptions = [
-	{ label: 'Normal', value: 'mix-blend-normal' },
-	{ label: 'Multiply', value: 'mix-blend-multiply' },
-	{ label: 'Screen', value: 'mix-blend-screen' },
-	{ label: 'Overlay', value: 'mix-blend-overlay' },
-	{ label: 'Darken', value: 'mix-blend-darken' },
-	{ label: 'Lighten', value: 'mix-blend-lighten' },
-	{ label: 'Color Dodge', value: 'mix-blend-color-dodge' },
-	{ label: 'Color Burn', value: 'mix-blend-color-burn' },
-	{ label: 'Hard Light', value: 'mix-blend-hard-light' },
-	{ label: 'Soft Light', value: 'mix-blend-soft-light' },
-	{ label: 'Difference', value: 'mix-blend-difference' },
-	{ label: 'Exclusion', value: 'mix-blend-exclusion' },
-	{ label: 'Hue', value: 'mix-blend-hue' },
-	{ label: 'Saturation', value: 'mix-blend-saturation' },
-	{ label: 'Color', value: 'mix-blend-color' },
-	{ label: 'Luminosity', value: 'mix-blend-luminosity' },
 ];
 
 export const columnOptionsCore = [
