+++ data/bootstrap-classes/display-options.js
@@ -1,0 +1,68 @@
+export const displayOptions = [
+	{ label: 'Hide', value: 'd-none' },
+	{ label: 'Block Display', value: 'd-block' },
+	{ label: 'Inline Display', value: 'd-inline' },
+	{ label: 'Inline Block', value: 'd-inline-block' },
+	{ label: 'Flex Display', value: 'd-flex' },
+	{ label: 'Inline Flex', value: 'd-inline-flex' },
+	{ label: 'Table Display', value: 'd-table' },
+	{ label: 'Table Row', value: 'd-table-row' },
+	{ label: 'Table Cell', value: 'd-table-cell' },
+	{ label: 'Grid Display', value: 'd-grid' },
+	{ label: 'Inline Grid', value: 'd-inline-grid' },
+	{ label: 'Hide on Small', value: 'd-sm-none' },
+	{ label: 'Block on Small', value: 'd-sm-block' },
+	{ label: 'Inline on Small', value: 'd-sm-inline' },
+	{ label: 'Inline Block on Small', value: 'd-sm-inline-block' },
+	{ label: 'Flex on Small', value: 'd-sm-flex' },
+	{ label: 'Inline Flex on Small', value: 'd-sm-inline-flex' },
+	{ label: 'Table on Small', value: 'd-sm-table' },
+	{ label: 'Table Row on Small', value: 'd-sm-table-row' },
+	{ label: 'Table Cell on Small', value: 'd-sm-table-cell' },
+	{ label: 'Grid on Small', value: 'd-sm-grid' },
+	{ label: 'Inline Grid on Small', value: 'd-sm-inline-grid' },
+	{ label: 'Hide on Medium', value: 'd-md-none' },
+	{ label: 'Block on Medium', value: 'd-md-block' },
+	{ label: 'Inline on Medium', value: 'd-md-inline' },
+	{ label: 'Inline Block on Medium', value: 'd-md-inline-block' },
+	{ label: 'Flex on Medium', value: 'd-md-flex' },
+	{ label: 'Inline Flex on Medium', value: 'd-md-inline-flex' },
+	{ label: 'Table on Medium', value: 'd-md-table' },
+	{ label: 'Table Row on Medium', value: 'd-md-table-row' },
+	{ label: 'Table Cell on Medium', value: 'd-md-table-cell' },
+	{ label: 'Grid on Medium', value: 'd-md-grid' },
+	{ label: 'Inline Grid on Medium', value: 'd-md-inline-grid' },
+	{ label: 'Hide on Large', value: 'd-lg-none' },
+	{ label: 'Block on Large', value: 'd-lg-block' },
+	{ label: 'Inline on Large', value: 'd-lg-inline' },
+	{ label: 'Inline Block on Large', value: 'd-lg-inline-block' },
+	{ label: 'Flex on Large', value: 'd-lg-flex' },
+	{ label: 'Inline Flex on Large', value: 'd-lg-inline-flex' },
+	{ label: 'Table on Large', value: 'd-lg-table' },
+	{ label: 'Table Row on Large', value: 'd-lg-table-row' },
+	{ label: 'Table Cell on Large', value: 'd-lg-table-cell' },
+	{ label: 'Grid on Large', value: 'd-lg-grid' },
+	{ label: 'Inline Grid on Large', value: 'd-lg-inline-grid' },
+	{ label: 'Hide on Extra Large', value: 'd-xl-none' },
+	{ label: 'Block on Extra Large', value: 'd-xl-block' },
+	{ label: 'Inline on Extra Large', value: 'd-xl-inline' },
+	{ label: 'Inline Block on Extra Large', value: 'd-xl-inline-block' },
+	{ label: 'Flex on Extra Large', value: 'd-xl-flex' },
+	{ label: 'Inline Flex on Extra Large', value: 'd-xl-inline-flex' },
+	{ label: 'Table on Extra Large', value: 'd-xl-table' },
+	{ label: 'Table Row on Extra Large', value: 'd-xl-table-row' },
+	{ label: 'Table Cell on Extra Large', value: 'd-xl-table-cell' },
+	{ label: 'Grid on Extra Large', value: 'd-xl-grid' },
+	{ label: 'Inline Grid on Extra Large', value: 'd-xl-inline-grid' },
+	{ label: 'Hide on XXL', value: 'd-xxl-none' },
+	{ label: 'Block on XXL', value: 'd-xxl-block' },
+	{ label: 'Inline on XXL', value: 'd-xxl-inline' },
+	{ label: 'Inline Block on XXL', value: 'd-xxl-inline-block' },
+	{ label: 'Flex on XXL', value: 'd-xxl-flex' },
+	{ label: 'Inline Flex on XXL', value: 'd-xxl-inline-flex' },
+	{ label: 'Table on XXL', value: 'd-xxl-table' },
+	{ label: 'Table Row on XXL', value: 'd-xxl-table-row' },
+	{ label: 'Table Cell on XXL', value: 'd-xxl-table-cell' },
+	{ label: 'Grid on XXL', value: 'd-xxl-grid' },
+	{ label: 'Inline Grid on XXL', value: 'd-xxl-inline-grid' },
+];
