+++ data/bootstrap-classes/gap-options.js
@@ -1,0 +1,29 @@
+export const gapOptions = [
+	{ label: 'No Gap', value: 'gap-0' },
+	{ label: 'Gap 1', value: 'gap-1' },
+	{ label: 'Gap 2', value: 'gap-2' },
+	{ label: 'Gap 3', value: 'gap-3' },
+	{ label: 'Gap 4', value: 'gap-4' },
+	{ label: 'Gap 5', value: 'gap-5' },
+	{ label: 'Small Screen No Gap', value: 'gap-sm-0' },
+	{ label: 'Small Screen Gap 1', value: 'gap-sm-1' },
+	{ label: 'Small Screen Gap 2', value: 'gap-sm-2' },
+	{ label: 'Small Screen Gap 3', value: 'gap-sm-3' },
+	{ label: 'Small Screen Gap 4', value: 'gap-sm-4' },
+	{ label: 'Small Screen Gap 5', value: 'gap-sm-5' },
+	{ label: 'Medium Screen No Gap', value: 'gap-md-0' },
+	{ label: 'Medium Screen Gap 1', value: 'gap-md-1' },
+	{ label: 'Medium Screen Gap 2', value: 'gap-md-2' },
+	{ label: 'Medium Screen Gap 3', value: 'gap-md-3' },
+	{ label: 'Medium Screen Gap 4', value: 'gap-md-4' },
+	{ label: 'Medium Screen Gap 5', value: 'gap-md-5' },
+	{ label: 'Large Screen No Gap', value: 'gap-lg-0' },
+	{ label: 'Large Screen Gap 1', value: 'gap-lg-1' },
+	{ label: 'Large Screen Gap 2', value: 'gap-lg-2' },
+	{ label: 'Large Screen Gap 3', value: 'gap-lg-3' },
+	{ label: 'Large Screen Gap 4', value: 'gap-lg-4' },
+	{ label: 'Large Screen Gap 5', value: 'gap-lg-5' },
+	{ label: 'Extra Large No Gap', value: 'gap-xl-0' },
+	{ label: 'Extra Large Gap 1', value: 'gap-xl-1' },
+	{ label: 'Extra Large Gap 2', value: 'gap-xl-2' },
+];
