+++ data/bootstrap-classes/justify-content-options.js
@@ -1,0 +1,37 @@
+export const justifyContentOptions = [
+	{ label: 'Justify Start', value: 'justify-content-start' },
+	{ label: 'Medium Screen Justify Start', value: 'justify-content-md-start' },
+	{ label: 'Large Screen Justify Start', value: 'justify-content-lg-start' },
+	{ label: 'Extra Large Justify Start', value: 'justify-content-xl-start' },
+	{ label: 'XXL Justify Start', value: 'justify-content-xxl-start' },
+	{ label: 'Justify End', value: 'justify-content-end' },
+	{ label: 'Medium Screen Justify End', value: 'justify-content-md-end' },
+	{ label: 'Large Screen Justify End', value: 'justify-content-lg-end' },
+	{ label: 'Extra Large Justify End', value: 'justify-content-xl-end' },
+	{ label: 'XXL Justify End', value: 'justify-content-xxl-end' },
+	{ label: 'Justify Center', value: 'justify-content-center' },
+	{
+		label: 'Medium Screen Justify Center',
+		value: 'justify-content-md-center',
+	},
+	{
+		label: 'Large Screen Justify Center',
+		value: 'justify-content-lg-center',
+	},
+	{ label: 'Extra Large Justify Center', value: 'justify-content-xl-center' },
+	{ label: 'XXL Justify Center', value: 'justify-content-xxl-center' },
+	{ label: 'Justify Between', value: 'justify-content-between' },
+	{
+		label: 'Medium Screen Justify Between',
+		value: 'justify-content-md-between',
+	},
+	{
+		label: 'Large Screen Justify Between',
+		value: 'justify-content-lg-between',
+	},
+	{
+		label: 'Extra Large Justify Between',
+		value: 'justify-content-xl-between',
+	},
+	{ label: 'XXL Justify Between', value: 'justify-content-xxl-between' },
+];
