+++ data/bootstrap-classes/margin-options.js
@@ -1,0 +1,156 @@
+export const marginOptions = [
+	{ label: 'Start Margin Auto', value: 'ms-auto' },
+	{ label: 'End Margin Auto', value: 'me-auto' },
+	{ label: 'Small Screen End Margin Auto', value: 'me-sm-auto' },
+	{ label: 'Medium Screen End Margin Auto', value: 'me-md-auto' },
+	{ label: 'Medium Screen Start Margin Auto', value: 'ms-md-auto' },
+	{ label: 'Large Screen End Margin Auto', value: 'me-lg-auto' },
+	{ label: 'Large Screen Start Margin Auto', value: 'ms-lg-auto' },
+	{ label: 'Extra Large End Margin Auto', value: 'me-xl-auto' },
+	{ label: 'Extra Large Start Margin Auto', value: 'ms-xl-auto' },
+	{ label: 'XXL End Margin Auto', value: 'me-xxl-auto' },
+	{ label: 'XXL Start Margin Auto', value: 'ms-xxl-auto' },
+	{ label: 'No Margin', value: 'm-0' },
+	{ label: 'Margin 1', value: 'm-1' },
+	{ label: 'Margin 2', value: 'm-2' },
+	{ label: 'Margin 3', value: 'm-3' },
+	{ label: 'Margin 4', value: 'm-4' },
+	{ label: 'Margin 5', value: 'm-5' },
+	{ label: 'Margin 6', value: 'm-6' },
+	{ label: 'Margin 7', value: 'm-7' },
+	{ label: 'Margin 8', value: 'm-8' },
+	{ label: 'Margin 9', value: 'm-9' },
+	{ label: 'Margin 10', value: 'm-10' },
+	{ label: 'Bottom Margin 0', value: 'mb-0' },
+	{ label: 'Bottom Margin 1', value: 'mb-1' },
+	{ label: 'Bottom Margin 2', value: 'mb-2' },
+	{ label: 'Bottom Margin 3', value: 'mb-3' },
+	{ label: 'Bottom Margin 4', value: 'mb-4' },
+	{ label: 'Bottom Margin 5', value: 'mb-5' },
+	{ label: 'Bottom Margin 6', value: 'mb-6' },
+	{ label: 'Bottom Margin 7', value: 'mb-7' },
+	{ label: 'Bottom Margin 8', value: 'mb-8' },
+	{ label: 'Bottom Margin 9', value: 'mb-9' },
+	{ label: 'Bottom Margin 10', value: 'mb-10' },
+	{ label: 'Top Margin 0', value: 'mt-0' },
+	{ label: 'Top Margin 1', value: 'mt-1' },
+	{ label: 'Top Margin 2', value: 'mt-2' },
+	{ label: 'Top Margin 3', value: 'mt-3' },
+	{ label: 'Top Margin 4', value: 'mt-4' },
+	{ label: 'Top Margin 5', value: 'mt-5' },
+	{ label: 'Top Margin 6', value: 'mt-6' },
+	{ label: 'Top Margin 7', value: 'mt-7' },
+	{ label: 'Top Margin 8', value: 'mt-8' },
+	{ label: 'Top Margin 9', value: 'mt-9' },
+	{ label: 'Top Margin 10', value: 'mt-10' },
+	{ label: 'Negative Start Margin 0', value: 'ms-n0' },
+	{ label: 'Negative Start Margin 1', value: 'ms-n1' },
+	{ label: 'Negative Start Margin 2', value: 'ms-n2' },
+	{ label: 'Negative Start Margin 3', value: 'ms-n3' },
+	{ label: 'Negative Start Margin 4', value: 'ms-n4' },
+	{ label: 'Negative Start Margin 5', value: 'ms-n5' },
+	{ label: 'Negative Start Margin 6', value: 'ms-n6' },
+	{ label: 'Negative Start Margin 7', value: 'ms-n7' },
+	{ label: 'Negative Start Margin 8', value: 'ms-n8' },
+	{ label: 'Negative Start Margin 9', value: 'ms-n9' },
+	{ label: 'Negative Start Margin 10', value: 'ms-n10' },
+	{ label: 'Negative End Margin 0', value: 'me-n0' },
+	{ label: 'Negative End Margin 1', value: 'me-n1' },
+	{ label: 'Negative End Margin 2', value: 'me-n2' },
+	{ label: 'Negative End Margin 3', value: 'me-n3' },
+	{ label: 'Negative End Margin 4', value: 'me-n4' },
+	{ label: 'Negative End Margin 5', value: 'me-n5' },
+	{ label: 'Negative End Margin 6', value: 'me-n6' },
+	{ label: 'Negative End Margin 7', value: 'me-n7' },
+	{ label: 'Negative End Margin 8', value: 'me-n8' },
+	{ label: 'Negative End Margin 9', value: 'me-n9' },
+	{ label: 'Negative End Margin 10', value: 'me-n10' },
+	{ label: 'Medium Negative Start Margin 0', value: 'ms-md-n0' },
+	{ label: 'Medium Negative Start Margin 1', value: 'ms-md-n1' },
+	{ label: 'Medium Negative Start Margin 2', value: 'ms-md-n2' },
+	{ label: 'Medium Negative Start Margin 3', value: 'ms-md-n3' },
+	{ label: 'Medium Negative Start Margin 4', value: 'ms-md-n4' },
+	{ label: 'Medium Negative Start Margin 5', value: 'ms-md-n5' },
+	{ label: 'Medium Negative Start Margin 6', value: 'ms-md-n6' },
+	{ label: 'Medium Negative Start Margin 7', value: 'ms-md-n7' },
+	{ label: 'Medium Negative Start Margin 8', value: 'ms-md-n8' },
+	{ label: 'Medium Negative Start Margin 9', value: 'ms-md-n9' },
+	{ label: 'Medium Negative Start Margin 10', value: 'ms-md-n10' },
+	{ label: 'Large Negative Start Margin 0', value: 'ms-lg-n0' },
+	{ label: 'Large Negative Start Margin 1', value: 'ms-lg-n1' },
+	{ label: 'Large Negative Start Margin 2', value: 'ms-lg-n2' },
+	{ label: 'Large Negative Start Margin 3', value: 'ms-lg-n3' },
+	{ label: 'Large Negative Start Margin 4', value: 'ms-lg-n4' },
+	{ label: 'Large Negative Start Margin 5', value: 'ms-lg-n5' },
+	{ label: 'Large Negative Start Margin 6', value: 'ms-lg-n6' },
+	{ label: 'Large Negative Start Margin 7', value: 'ms-lg-n7' },
+	{ label: 'Large Negative Start Margin 8', value: 'ms-lg-n8' },
+	{ label: 'Large Negative Start Margin 9', value: 'ms-lg-n9' },
+	{ label: 'Large Negative Start Margin 10', value: 'ms-lg-n10' },
+	{ label: 'Extra Large Negative Start Margin 0', value: 'ms-xl-n0' },
+	{ label: 'Extra Large Negative Start Margin 1', value: 'ms-xl-n1' },
+	{ label: 'Extra Large Negative Start Margin 2', value: 'ms-xl-n2' },
+	{ label: 'Extra Large Negative Start Margin 3', value: 'ms-xl-n3' },
+	{ label: 'Extra Large Negative Start Margin 4', value: 'ms-xl-n4' },
+	{ label: 'Extra Large Negative Start Margin 5', value: 'ms-xl-n5' },
+	{ label: 'Extra Large Negative Start Margin 6', value: 'ms-xl-n6' },
+	{ label: 'Extra Large Negative Start Margin 7', value: 'ms-xl-n7' },
+	{ label: 'Extra Large Negative Start Margin 8', value: 'ms-xl-n8' },
+	{ label: 'Extra Large Negative Start Margin 9', value: 'ms-xl-n9' },
+	{ label: 'Extra Large Negative Start Margin 10', value: 'ms-xl-n10' },
+	{ label: 'XXL Negative Start Margin 0', value: 'ms-xxl-n0' },
+	{ label: 'XXL Negative Start Margin 1', value: 'ms-xxl-n1' },
+	{ label: 'XXL Negative Start Margin 2', value: 'ms-xxl-n2' },
+	{ label: 'XXL Negative Start Margin 3', value: 'ms-xxl-n3' },
+	{ label: 'XXL Negative Start Margin 4', value: 'ms-xxl-n4' },
+	{ label: 'XXL Negative Start Margin 5', value: 'ms-xxl-n5' },
+	{ label: 'XXL Negative Start Margin 6', value: 'ms-xxl-n6' },
+	{ label: 'XXL Negative Start Margin 7', value: 'ms-xxl-n7' },
+	{ label: 'XXL Negative Start Margin 8', value: 'ms-xxl-n8' },
+	{ label: 'XXL Negative Start Margin 9', value: 'ms-xxl-n9' },
+	{ label: 'XXL Negative Start Margin 10', value: 'ms-xxl-n10' },
+	{ label: 'Medium Negative End Margin 0', value: 'me-md-n0' },
+	{ label: 'Medium Negative End Margin 1', value: 'me-md-n1' },
+	{ label: 'Medium Negative End Margin 2', value: 'me-md-n2' },
+	{ label: 'Medium Negative End Margin 3', value: 'me-md-n3' },
+	{ label: 'Medium Negative End Margin 4', value: 'me-md-n4' },
+	{ label: 'Medium Negative End Margin 5', value: 'me-md-n5' },
+	{ label: 'Medium Negative End Margin 6', value: 'me-md-n6' },
+	{ label: 'Medium Negative End Margin 7', value: 'me-md-n7' },
+	{ label: 'Medium Negative End Margin 8', value: 'me-md-n8' },
+	{ label: 'Medium Negative End Margin 9', value: 'me-md-n9' },
+	{ label: 'Medium Negative End Margin 10', value: 'me-md-n10' },
+	{ label: 'Large Negative End Margin 0', value: 'me-lg-n0' },
+	{ label: 'Large Negative End Margin 1', value: 'me-lg-n1' },
+	{ label: 'Large Negative End Margin 2', value: 'me-lg-n2' },
+	{ label: 'Large Negative End Margin 3', value: 'me-lg-n3' },
+	{ label: 'Large Negative End Margin 4', value: 'me-lg-n4' },
+	{ label: 'Large Negative End Margin 5', value: 'me-lg-n5' },
+	{ label: 'Large Negative End Margin 6', value: 'me-lg-n6' },
+	{ label: 'Large Negative End Margin 7', value: 'me-lg-n7' },
+	{ label: 'Large Negative End Margin 8', value: 'me-lg-n8' },
+	{ label: 'Large Negative End Margin 9', value: 'me-lg-n9' },
+	{ label: 'Large Negative End Margin 10', value: 'me-lg-n10' },
+	{ label: 'Extra Large Negative End Margin 0', value: 'me-xl-n0' },
+	{ label: 'Extra Large Negative End Margin 1', value: 'me-xl-n1' },
+	{ label: 'Extra Large Negative End Margin 2', value: 'me-xl-n2' },
+	{ label: 'Extra Large Negative End Margin 3', value: 'me-xl-n3' },
+	{ label: 'Extra Large Negative End Margin 4', value: 'me-xl-n4' },
+	{ label: 'Extra Large Negative End Margin 5', value: 'me-xl-n5' },
+	{ label: 'Extra Large Negative End Margin 6', value: 'me-xl-n6' },
+	{ label: 'Extra Large Negative End Margin 7', value: 'me-xl-n7' },
+	{ label: 'Extra Large Negative End Margin 8', value: 'me-xl-n8' },
+	{ label: 'Extra Large Negative End Margin 9', value: 'me-xl-n9' },
+	{ label: 'Extra Large Negative End Margin 10', value: 'me-xl-n10' },
+	{ label: 'XXL Negative End Margin 0', value: 'me-xxl-n0' },
+	{ label: 'XXL Negative End Margin 1', value: 'me-xxl-n1' },
+	{ label: 'XXL Negative End Margin 2', value: 'me-xxl-n2' },
+	{ label: 'XXL Negative End Margin 3', value: 'me-xxl-n3' },
+	{ label: 'XXL Negative End Margin 4', value: 'me-xxl-n4' },
+	{ label: 'XXL Negative End Margin 5', value: 'me-xxl-n5' },
+	{ label: 'XXL Negative End Margin 6', value: 'me-xxl-n6' },
+	{ label: 'XXL Negative End Margin 7', value: 'me-xxl-n7' },
+	{ label: 'XXL Negative End Margin 8', value: 'me-xxl-n8' },
+	{ label: 'XXL Negative End Margin 9', value: 'me-xxl-n9' },
+	{ label: 'XXL Negative End Margin 10', value: 'me-xxl-n10' },
+];
