+++ data/bootstrap-classes/order-options.js
@@ -1,0 +1,50 @@
+export const orderOptions = [
+	{ label: 'Order 0', value: 'order-0' },
+	{ label: 'Order 1', value: 'order-1' },
+	{ label: 'Order 2', value: 'order-2' },
+	{ label: 'Order 3', value: 'order-3' },
+	{ label: 'Order 4', value: 'order-4' },
+	{ label: 'Order 5', value: 'order-5' },
+	{ label: 'Small Screen Order 0', value: 'order-sm-0' },
+	{ label: 'Small Screen Order 1', value: 'order-sm-1' },
+	{ label: 'Small Screen Order 2', value: 'order-sm-2' },
+	{ label: 'Small Screen Order 3', value: 'order-sm-3' },
+	{ label: 'Small Screen Order 4', value: 'order-sm-4' },
+	{ label: 'Small Screen Order 5', value: 'order-sm-5' },
+	{ label: 'Medium Screen Order 0', value: 'order-md-0' },
+	{ label: 'Medium Screen Order 1', value: 'order-md-1' },
+	{ label: 'Medium Screen Order 2', value: 'order-md-2' },
+	{ label: 'Medium Screen Order 3', value: 'order-md-3' },
+	{ label: 'Medium Screen Order 4', value: 'order-md-4' },
+	{ label: 'Medium Screen Order 5', value: 'order-md-5' },
+	{ label: 'Large Screen Order 0', value: 'order-lg-0' },
+	{ label: 'Large Screen Order 1', value: 'order-lg-1' },
+	{ label: 'Large Screen Order 2', value: 'order-lg-2' },
+	{ label: 'Large Screen Order 3', value: 'order-lg-3' },
+	{ label: 'Large Screen Order 4', value: 'order-lg-4' },
+	{ label: 'Large Screen Order 5', value: 'order-lg-5' },
+	{ label: 'Extra Large Order 0', value: 'order-xl-0' },
+	{ label: 'Extra Large Order 1', value: 'order-xl-1' },
+	{ label: 'Extra Large Order 2', value: 'order-xl-2' },
+	{ label: 'Extra Large Order 3', value: 'order-xl-3' },
+	{ label: 'Extra Large Order 4', value: 'order-xl-4' },
+	{ label: 'Extra Large Order 5', value: 'order-xl-5' },
+	{ label: 'XXL Order 0', value: 'order-xxl-0' },
+	{ label: 'XXL Order 1', value: 'order-xxl-1' },
+	{ label: 'XXL Order 2', value: 'order-xxl-2' },
+	{ label: 'XXL Order 3', value: 'order-xxl-3' },
+	{ label: 'XXL Order 4', value: 'order-xxl-4' },
+	{ label: 'XXL Order 5', value: 'order-xxl-5' },
+	{ label: 'Order First', value: 'order-first' },
+	{ label: 'Small Screen Order First', value: 'order-sm-first' },
+	{ label: 'Medium Screen Order First', value: 'order-md-first' },
+	{ label: 'Large Screen Order First', value: 'order-lg-first' },
+	{ label: 'Extra Large Order First', value: 'order-xl-first' },
+	{ label: 'XXL Order First', value: 'order-xxl-first' },
+	{ label: 'Order Last', value: 'order-last' },
+	{ label: 'Small Screen Order Last', value: 'order-sm-last' },
+	{ label: 'Medium Screen Order Last', value: 'order-md-last' },
+	{ label: 'Large Screen Order Last', value: 'order-lg-last' },
+	{ label: 'Extra Large Order Last', value: 'order-xl-last' },
+	{ label: 'XXL Order Last', value: 'order-xxl-last' },
+];
