+++ data/bootstrap-classes/padding-options.js
@@ -1,0 +1,200 @@
+export const paddingOptions = [
+	{ label: 'No Padding', value: 'p-0' },
+	{ label: 'Padding 1', value: 'p-1' },
+	{ label: 'Padding 2', value: 'p-2' },
+	{ label: 'Padding 3', value: 'p-3' },
+	{ label: 'Padding 4', value: 'p-4' },
+	{ label: 'Padding 5', value: 'p-5' },
+	{ label: 'Padding 6', value: 'p-6' },
+	{ label: 'Padding 7', value: 'p-7' },
+	{ label: 'Padding 8', value: 'p-8' },
+	{ label: 'Padding 9', value: 'p-9' },
+	{ label: 'Padding 10', value: 'p-10' },
+	{ label: 'Small Screen No Padding', value: 'p-sm-0' },
+	{ label: 'Small Screen Padding 1', value: 'p-sm-1' },
+	{ label: 'Small Screen Padding 2', value: 'p-sm-2' },
+	{ label: 'Small Screen Padding 3', value: 'p-sm-3' },
+	{ label: 'Small Screen Padding 4', value: 'p-sm-4' },
+	{ label: 'Small Screen Padding 5', value: 'p-sm-5' },
+	{ label: 'Small Screen Padding 6', value: 'p-sm-6' },
+	{ label: 'Small Screen Padding 7', value: 'p-sm-7' },
+	{ label: 'Small Screen Padding 8', value: 'p-sm-8' },
+	{ label: 'Small Screen Padding 9', value: 'p-sm-9' },
+	{ label: 'Small Screen Padding 10', value: 'p-sm-10' },
+	{ label: 'Medium Screen No Padding', value: 'p-md-0' },
+	{ label: 'Medium Screen Padding 1', value: 'p-md-1' },
+	{ label: 'Medium Screen Padding 2', value: 'p-md-2' },
+	{ label: 'Medium Screen Padding 3', value: 'p-md-3' },
+	{ label: 'Medium Screen Padding 4', value: 'p-md-4' },
+	{ label: 'Medium Screen Padding 5', value: 'p-md-5' },
+	{ label: 'Medium Screen Padding 6', value: 'p-md-6' },
+	{ label: 'Medium Screen Padding 7', value: 'p-md-7' },
+	{ label: 'Medium Screen Padding 8', value: 'p-md-8' },
+	{ label: 'Medium Screen Padding 9', value: 'p-md-9' },
+	{ label: 'Medium Screen Padding 10', value: 'p-md-10' },
+	{ label: 'Large Screen No Padding', value: 'p-lg-0' },
+	{ label: 'Large Screen Padding 1', value: 'p-lg-1' },
+	{ label: 'Large Screen Padding 2', value: 'p-lg-2' },
+	{ label: 'Large Screen Padding 3', value: 'p-lg-3' },
+	{ label: 'Large Screen Padding 4', value: 'p-lg-4' },
+	{ label: 'Large Screen Padding 5', value: 'p-lg-5' },
+	{ label: 'Large Screen Padding 6', value: 'p-lg-6' },
+	{ label: 'Large Screen Padding 7', value: 'p-lg-7' },
+	{ label: 'Large Screen Padding 8', value: 'p-lg-8' },
+	{ label: 'Large Screen Padding 9', value: 'p-lg-9' },
+	{ label: 'Large Screen Padding 10', value: 'p-lg-10' },
+	{ label: 'Extra Large No Padding', value: 'p-xl-0' },
+	{ label: 'Extra Large Padding 1', value: 'p-xl-1' },
+	{ label: 'Extra Large Padding 2', value: 'p-xl-2' },
+	{ label: 'Extra Large Padding 3', value: 'p-xl-3' },
+	{ label: 'Extra Large Padding 4', value: 'p-xl-4' },
+	{ label: 'Extra Large Padding 5', value: 'p-xl-5' },
+	{ label: 'Extra Large Padding 6', value: 'p-xl-6' },
+	{ label: 'Extra Large Padding 7', value: 'p-xl-7' },
+	{ label: 'Extra Large Padding 8', value: 'p-xl-8' },
+	{ label: 'Extra Large Padding 9', value: 'p-xl-9' },
+	{ label: 'Extra Large Padding 10', value: 'p-xl-10' },
+	{ label: 'XXL No Padding', value: 'p-xxl-0' },
+	{ label: 'XXL Padding 1', value: 'p-xxl-1' },
+	{ label: 'XXL Padding 2', value: 'p-xxl-2' },
+	{ label: 'XXL Padding 3', value: 'p-xxl-3' },
+	{ label: 'XXL Padding 4', value: 'p-xxl-4' },
+	{ label: 'XXL Padding 5', value: 'p-xxl-5' },
+	{ label: 'XXL Padding 6', value: 'p-xxl-6' },
+	{ label: 'XXL Padding 7', value: 'p-xxl-7' },
+	{ label: 'XXL Padding 8', value: 'p-xxl-8' },
+	{ label: 'XXL Padding 9', value: 'p-xxl-9' },
+	{ label: 'XXL Padding 10', value: 'p-xxl-10' },
+	{ label: 'Top Padding 0', value: 'pt-0' },
+	{ label: 'Top Padding 1', value: 'pt-1' },
+	{ label: 'Top Padding 2', value: 'pt-2' },
+	{ label: 'Top Padding 3', value: 'pt-3' },
+	{ label: 'Top Padding 4', value: 'pt-4' },
+	{ label: 'Top Padding 5', value: 'pt-5' },
+	{ label: 'Top Padding 6', value: 'pt-6' },
+	{ label: 'Top Padding 7', value: 'pt-7' },
+	{ label: 'Top Padding 8', value: 'pt-8' },
+	{ label: 'Top Padding 9', value: 'pt-9' },
+	{ label: 'Top Padding 10', value: 'pt-10' },
+	{ label: 'Small Screen Top Padding 0', value: 'pt-sm-0' },
+	{ label: 'Small Screen Top Padding 1', value: 'pt-sm-1' },
+	{ label: 'Small Screen Top Padding 2', value: 'pt-sm-2' },
+	{ label: 'Small Screen Top Padding 3', value: 'pt-sm-3' },
+	{ label: 'Small Screen Top Padding 4', value: 'pt-sm-4' },
+	{ label: 'Small Screen Top Padding 5', value: 'pt-sm-5' },
+	{ label: 'Small Screen Top Padding 6', value: 'pt-sm-6' },
+	{ label: 'Small Screen Top Padding 7', value: 'pt-sm-7' },
+	{ label: 'Small Screen Top Padding 8', value: 'pt-sm-8' },
+	{ label: 'Small Screen Top Padding 9', value: 'pt-sm-9' },
+	{ label: 'Small Screen Top Padding 10', value: 'pt-sm-10' },
+	{ label: 'Medium Screen Top Padding 0', value: 'pt-md-0' },
+	{ label: 'Medium Screen Top Padding 1', value: 'pt-md-1' },
+	{ label: 'Medium Screen Top Padding 2', value: 'pt-md-2' },
+	{ label: 'Medium Screen Top Padding 3', value: 'pt-md-3' },
+	{ label: 'Medium Screen Top Padding 4', value: 'pt-md-4' },
+	{ label: 'Medium Screen Top Padding 5', value: 'pt-md-5' },
+	{ label: 'Medium Screen Top Padding 6', value: 'pt-md-6' },
+	{ label: 'Medium Screen Top Padding 7', value: 'pt-md-7' },
+	{ label: 'Medium Screen Top Padding 8', value: 'pt-md-8' },
+	{ label: 'Medium Screen Top Padding 9', value: 'pt-md-9' },
+	{ label: 'Medium Screen Top Padding 10', value: 'pt-md-10' },
+	{ label: 'Large Screen Top Padding 0', value: 'pt-lg-0' },
+	{ label: 'Large Screen Top Padding 1', value: 'pt-lg-1' },
+	{ label: 'Large Screen Top Padding 2', value: 'pt-lg-2' },
+	{ label: 'Large Screen Top Padding 3', value: 'pt-lg-3' },
+	{ label: 'Large Screen Top Padding 4', value: 'pt-lg-4' },
+	{ label: 'Large Screen Top Padding 5', value: 'pt-lg-5' },
+	{ label: 'Large Screen Top Padding 6', value: 'pt-lg-6' },
+	{ label: 'Large Screen Top Padding 7', value: 'pt-lg-7' },
+	{ label: 'Large Screen Top Padding 8', value: 'pt-lg-8' },
+	{ label: 'Large Screen Top Padding 9', value: 'pt-lg-9' },
+	{ label: 'Large Screen Top Padding 10', value: 'pt-lg-10' },
+	{ label: 'Extra Large Top Padding 0', value: 'pt-xl-0' },
+	{ label: 'Extra Large Top Padding 1', value: 'pt-xl-1' },
+	{ label: 'Extra Large Top Padding 2', value: 'pt-xl-2' },
+	{ label: 'Extra Large Top Padding 3', value: 'pt-xl-3' },
+	{ label: 'Extra Large Top Padding 4', value: 'pt-xl-4' },
+	{ label: 'Extra Large Top Padding 5', value: 'pt-xl-5' },
+	{ label: 'Extra Large Top Padding 6', value: 'pt-xl-6' },
+	{ label: 'Extra Large Top Padding 7', value: 'pt-xl-7' },
+	{ label: 'Extra Large Top Padding 8', value: 'pt-xl-8' },
+	{ label: 'Extra Large Top Padding 9', value: 'pt-xl-9' },
+	{ label: 'Extra Large Top Padding 10', value: 'pt-xl-10' },
+	{ label: 'XXL Top Padding 0', value: 'pt-xxl-0' },
+	{ label: 'XXL Top Padding 1', value: 'pt-xxl-1' },
+	{ label: 'XXL Top Padding 2', value: 'pt-xxl-2' },
+	{ label: 'XXL Top Padding 3', value: 'pt-xxl-3' },
+	{ label: 'XXL Top Padding 4', value: 'pt-xxl-4' },
+	{ label: 'XXL Top Padding 5', value: 'pt-xxl-5' },
+	{ label: 'XXL Top Padding 6', value: 'pt-xxl-6' },
+	{ label: 'XXL Top Padding 7', value: 'pt-xxl-7' },
+	{ label: 'XXL Top Padding 8', value: 'pt-xxl-8' },
+	{ label: 'XXL Top Padding 9', value: 'pt-xxl-9' },
+	{ label: 'XXL Top Padding 10', value: 'pt-xxl-10' },
+	{ label: 'Bottom Padding 0', value: 'pb-0' },
+	{ label: 'Bottom Padding 1', value: 'pb-1' },
+	{ label: 'Bottom Padding 2', value: 'pb-2' },
+	{ label: 'Bottom Padding 3', value: 'pb-3' },
+	{ label: 'Bottom Padding 4', value: 'pb-4' },
+	{ label: 'Bottom Padding 5', value: 'pb-5' },
+	{ label: 'Bottom Padding 6', value: 'pb-6' },
+	{ label: 'Bottom Padding 7', value: 'pb-7' },
+	{ label: 'Bottom Padding 8', value: 'pb-8' },
+	{ label: 'Bottom Padding 9', value: 'pb-9' },
+	{ label: 'Bottom Padding 10', value: 'pb-10' },
+	{ label: 'Small Screen Bottom Padding 0', value: 'pb-sm-0' },
+	{ label: 'Small Screen Bottom Padding 1', value: 'pb-sm-1' },
+	{ label: 'Small Screen Bottom Padding 2', value: 'pb-sm-2' },
+	{ label: 'Small Screen Bottom Padding 3', value: 'pb-sm-3' },
+	{ label: 'Small Screen Bottom Padding 4', value: 'pb-sm-4' },
+	{ label: 'Small Screen Bottom Padding 5', value: 'pb-sm-5' },
+	{ label: 'Small Screen Bottom Padding 6', value: 'pb-sm-6' },
+	{ label: 'Small Screen Bottom Padding 7', value: 'pb-sm-7' },
+	{ label: 'Small Screen Bottom Padding 8', value: 'pb-sm-8' },
+	{ label: 'Small Screen Bottom Padding 9', value: 'pb-sm-9' },
+	{ label: 'Small Screen Bottom Padding 10', value: 'pb-sm-10' },
+	{ label: 'Medium Screen Bottom Padding 0', value: 'pb-md-0' },
+	{ label: 'Medium Screen Bottom Padding 1', value: 'pb-md-1' },
+	{ label: 'Medium Screen Bottom Padding 2', value: 'pb-md-2' },
+	{ label: 'Medium Screen Bottom Padding 3', value: 'pb-md-3' },
+	{ label: 'Medium Screen Bottom Padding 4', value: 'pb-md-4' },
+	{ label: 'Medium Screen Bottom Padding 5', value: 'pb-md-5' },
+	{ label: 'Medium Screen Bottom Padding 6', value: 'pb-md-6' },
+	{ label: 'Medium Screen Bottom Padding 7', value: 'pb-md-7' },
+	{ label: 'Medium Screen Bottom Padding 8', value: 'pb-md-8' },
+	{ label: 'Medium Screen Bottom Padding 9', value: 'pb-md-9' },
+	{ label: 'Medium Screen Bottom Padding 10', value: 'pb-md-10' },
+	{ label: 'Large Screen Bottom Padding 0', value: 'pb-lg-0' },
+	{ label: 'Large Screen Bottom Padding 1', value: 'pb-lg-1' },
+	{ label: 'Large Screen Bottom Padding 2', value: 'pb-lg-2' },
+	{ label: 'Large Screen Bottom Padding 3', value: 'pb-lg-3' },
+	{ label: 'Large Screen Bottom Padding 4', value: 'pb-lg-4' },
+	{ label: 'Large Screen Bottom Padding 5', value: 'pb-lg-5' },
+	{ label: 'Large Screen Bottom Padding 6', value: 'pb-lg-6' },
+	{ label: 'Large Screen Bottom Padding 7', value: 'pb-lg-7' },
+	{ label: 'Large Screen Bottom Padding 8', value: 'pb-lg-8' },
+	{ label: 'Large Screen Bottom Padding 9', value: 'pb-lg-9' },
+	{ label: 'Large Screen Bottom Padding 10', value: 'pb-lg-10' },
+	{ label: 'Extra Large Bottom Padding 0', value: 'pb-xl-0' },
+	{ label: 'Extra Large Bottom Padding 1', value: 'pb-xl-1' },
+	{ label: 'Extra Large Bottom Padding 2', value: 'pb-xl-2' },
+	{ label: 'Extra Large Bottom Padding 3', value: 'pb-xl-3' },
+	{ label: 'Extra Large Bottom Padding 4', value: 'pb-xl-4' },
+	{ label: 'Extra Large Bottom Padding 5', value: 'pb-xl-5' },
+	{ label: 'Extra Large Bottom Padding 6', value: 'pb-xl-6' },
+	{ label: 'Extra Large Bottom Padding 7', value: 'pb-xl-7' },
+	{ label: 'Extra Large Bottom Padding 8', value: 'pb-xl-8' },
+	{ label: 'Extra Large Bottom Padding 9', value: 'pb-xl-9' },
+	{ label: 'Extra Large Bottom Padding 10', value: 'pb-xl-10' },
+	{ label: 'XXL Bottom Padding 0', value: 'pb-xxl-0' },
+	{ label: 'XXL Bottom Padding 1', value: 'pb-xxl-1' },
+	{ label: 'XXL Bottom Padding 2', value: 'pb-xxl-2' },
+	{ label: 'XXL Bottom Padding 3', value: 'pb-xxl-3' },
+	{ label: 'XXL Bottom Padding 4', value: 'pb-xxl-4' },
+	{ label: 'XXL Bottom Padding 5', value: 'pb-xxl-5' },
+	{ label: 'XXL Bottom Padding 6', value: 'pb-xxl-6' },
+	{ label: 'XXL Bottom Padding 7', value: 'pb-xxl-7' },
+	{ label: 'XXL Bottom Padding 8', value: 'pb-xxl-8' },
+	{ label: 'XXL Bottom Padding 9', value: 'pb-xxl-9' },
+	{ label: 'XXL Bottom Padding 10', value: 'pb-xxl-10' },
+];
