+++ data/bootstrap-classes/position-options.js
@@ -1,0 +1,21 @@
+export const positionOptions = [
+	{ label: 'Static Position', value: 'position-static' },
+	{ label: 'Relative Position', value: 'position-relative' },
+	{ label: 'Absolute Position', value: 'position-absolute' },
+	{ label: 'Fixed Position', value: 'position-fixed' },
+	{ label: 'Sticky Position', value: 'position-sticky' },
+	{ label: 'Top 0%', value: 'top-0' },
+	{ label: 'Top 50%', value: 'top-50' },
+	{ label: 'Top 100%', value: 'top-100' },
+	{ label: 'Bottom 0%', value: 'bottom-0' },
+	{ label: 'Bottom 50%', value: 'bottom-50' },
+	{ label: 'Bottom 100%', value: 'bottom-100' },
+	{ label: 'Start 0%', value: 'start-0' },
+	{ label: 'Start 50%', value: 'start-50' },
+	{ label: 'Start 100%', value: 'start-100' },
+	{ label: 'End 0%', value: 'end-0' },
+	{ label: 'End 50%', value: 'end-50' },
+	{ label: 'End 100%', value: 'end-100' },
+	{ label: 'Translate Middle X', value: 'translate-middle-x' },
+	{ label: 'Translate Middle Y', value: 'translate-middle-y' },
+];
