+++ data/bootstrap-classes/self-alignment-options.js
@@ -1,0 +1,27 @@
+export const selfAlignmentOptions = [
+	{ label: 'Align Top', value: 'align-self-start' },
+	{ label: 'Medium Screen Align Top', value: 'align-self-md-start' },
+	{ label: 'Large Screen Align Top', value: 'align-self-lg-start' },
+	{ label: 'Extra Large Align Top', value: 'align-self-xl-start' },
+	{ label: 'XXL Align Top', value: 'align-self-xxl-start' },
+	{ label: 'Align Center', value: 'align-self-center' },
+	{ label: 'Medium Screen Align Center', value: 'align-self-md-center' },
+	{ label: 'Large Screen Align Center', value: 'align-self-lg-center' },
+	{ label: 'Extra Large Align Center', value: 'align-self-xl-center' },
+	{ label: 'XXL Align Center', value: 'align-self-xxl-center' },
+	{ label: 'Align Bottom', value: 'align-self-end' },
+	{ label: 'Medium Screen Align Bottom', value: 'align-self-md-end' },
+	{ label: 'Large Screen Align Bottom', value: 'align-self-lg-end' },
+	{ label: 'Extra Large Align Bottom', value: 'align-self-xl-end' },
+	{ label: 'XXL Align Bottom', value: 'align-self-xxl-end' },
+	{ label: 'Align Baseline', value: 'align-self-baseline' },
+	{ label: 'Medium Screen Align Baseline', value: 'align-self-md-baseline' },
+	{ label: 'Large Screen Align Baseline', value: 'align-self-lg-baseline' },
+	{ label: 'Extra Large Align Baseline', value: 'align-self-xl-baseline' },
+	{ label: 'XXL Align Baseline', value: 'align-self-xxl-baseline' },
+	{ label: 'Stretch Align', value: 'align-self-stretch' },
+	{ label: 'Medium Screen Stretch', value: 'align-self-md-stretch' },
+	{ label: 'Large Screen Stretch', value: 'align-self-lg-stretch' },
+	{ label: 'Extra Large Stretch', value: 'align-self-xl-stretch' },
+	{ label: 'XXL Stretch', value: 'align-self-xxl-stretch' },
+];
