+++ data/bootstrap-classes/zindex-options.js
@@ -1,0 +1,7 @@
+export const zindexOptions = [
+	{ label: 'Z-Index 0', value: 'z-0' },
+	{ label: 'Z-Index 1', value: 'z-1' },
+	{ label: 'Z-Index 2', value: 'z-2' },
+	{ label: 'Z-Index 3', value: 'z-3' },
+	{ label: 'Z-Index -1', value: 'z-n1' },
+];
